Fix bug in array bit ordering (#388)
The `ir_fn_to_boolector` function in the boolector-based verifier is
flattening the elements in the reverse order. This isn't consistent with
the verilog semantics and other parts of the codebase. Also, the
`flatten_bv` function isn't handling arrays. This pull request adds the
test case and fixes the bug.

Built on top of #387.