Fix fuzzer (#428)
The git spr seems not robust enough across changes in its configuration
and this caused some commits to be merged incorrectly and some PRs are
lost (including this one). This pull request rescues the commit.

Original commit message:

The fuzzer generates invalid bit_slice operations where start + width is
greater than the bit-width of the operand. This pull request rules out
this.

commit-id:1b3c18d7