Add prove-quickcheck command (#451)
It seems that there is no need to handle implicit tokens in QuickCheck,
as such a function will be translated into two functions,
the first with implicit token and the second generates a token and
supply to the first function. We can just prove on the second function.

commit-id:9ddd6534

---

**Stack**:
- #454
- #451 ⬅


⚠️ *Part of a stack created by [spr](https://github.com/ejoffe/spr). Do
not merge manually using the UI - doing so may have unexpected results.*