Make prove-quickcheck more robust by falling back to itok mangled versions when the wrapper isn't available
In the original XLS system, it seems that each quickcheck function will be individually translated by converting to an individual module with this function as the entry point. This will trigger the system to generate a wrapper (likely by `WrapEntryIfImplicitToken`) that fixes the activation bit to 1.

In our implementation, we do not convert into single function package. Then, our quickcheck functions may not get the wrapper functions. Thus, the lookup of the ordinary function name will fail.

The fix now lookup both the ones with or without implicit token and use our mechanism of fixing such bits.