Add parallel scheduler (#483)
This pull request adds a parallel scheduler for the proofs.

Tasks can be organized in a tree where each group of tasks may have the
`all`, `any`, or `first` semantics.

For the `all` semantics, all the subtrees must be successful to make the
overall task successful. A single failure will cause the scheduler to
shutdown all the tasks in the group and report.

For the `any` semantics, any success in the subtrees could lead to
overall task success. A single success will cause the scheduler to
shutdown all the tasks in the group.

For the `first` semantics, the group's result is determined by the first
finished task. All other tasks will be canceled after this.

The reason why we have a tree structure is that if we want to introduce
a tactic/proof strategy/decomposition system, the proof obligations will
be organized as a tree.