Refactor quickcheck prover interface, add missing flags (#609)
This pull request is a major refactoring for the QuickCheck interface
provided by `Prover`. Specifically, we moved the logic for resolving the
DSLX from the driver to the QuickCheck prover interface so that we can
unify the behavior for the toolchain (which only accepts DSLX) and the
native provers. This could greatly simplify the driver code.

Also, this PR adds some missing functionalities like test filters and
enum domain in bound flags to quickcheck prover.