Add tests to verify generated FuzzSamples have interesting properties. (#626)
The tests generate a bunch of samples and verify that some non-trivial
interesting samples are generated. This includes a minimal and maximal
sized samples, verifying a variety of ops are generated, different
number of parameters are generated, and graphs whose live subset are
interesting.

Other changes:
* Fix issue where FuzzSampleSameTypedPair was generating uninteresting
rhs graphs.
* Fixing the fuzz pair bug exposed a minor issue with the ECO algorithm
and the ECO algorithm fuzzer which are also fixed.
* Fix some type inference bugs in FuzzSample generation.
* Fix some warnings about unused Results.
* Change remap_payload_with to take a enumerate-style (usize, NodeRef)
function in which the first argument is the operand slot.