DSLX specializer (#638)
This pull request adds a dslx-specialize command for DSLX code. Given a
top function in a module, it emits a monomorphized module with all the
function specializations reachable from the top function.

This PR also changed our management of lifetime for `InterpValue` and
`ParametricEnv`. Unlike those AST nodes, these values, including the
newly introduced `InvocationCalleeData` etc., are not directly owned by
a module. They may or may not have a owner. To simplify the management
of them, we decide to manage their lifetime fully in Rust. These values
will always be cloned to have a fresh lifetime when generated. This may
cause a small performance overhead but would make their lifetime much
cleaner.